% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.d.heat.eq.dirichlet.R
\name{one.d.heat.eq.dirichlet}
\alias{one.d.heat.eq.dirichlet}
\title{Numerical Solution to the Heat Equation with Dirichlet Boundary Conditions}
\usage{
one.d.heat.eq.dirichlet(n, soln, alpha, dt, eps)
}
\arguments{
\item{n}{}

\item{soln}{}

\item{alpha}{}

\item{dt}{}

\item{eps}{}
}
\value{

}
\description{
Numerical Solution to the Heat Equation with Dirichlet Boundary Conditions
}
