% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heat2.R
\name{plot.heat2}
\alias{plot.heat2}
\title{Plotting for Solutions of the 2D Heat Equation}
\usage{
\method{plot}{heat2}(heat, ...)
}
\arguments{
\item{heat}{the object of class "heat2".}

\item{...}{any arguments to pass to both calls of image.}
}
\description{
Plotting for Solutions of the 2D Heat Equation
}
\examples{
plot(heateq2d(matrix(runif(10000), nrow = 100)))
}
