% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.d.heat.eq.R
\name{two.d.heat.eq}
\alias{two.d.heat.eq}
\title{Solution of the 2D Heat Equation}
\usage{
two.d.heat.eq(n = 1000, init.dat, c = 1, dt = 0.01, dx = 0.1,
  boundary = "neumann")
}
\arguments{
\item{n}{the number of time steps.}

\item{init.dat}{a matrix of initial values in two dimensions.}

\item{c}{a conductivity constant.}

\item{dt}{the length of the time step.}

\item{dx}{the length of the spatial steps in both directions.}

\item{boundary}{boundary condition, either "neumann" or "dirichlet".}
}
\value{
The return value is a list of type "heat2" that includes the initial condition, final result (a matrix), and the boundary condition.
}
\description{
Given the formula dH/dt = c*(d^2H/dx^2+d^2H/dy^2), we can assign a meaning to all of the parameters.
}
\examples{
two.d.heat.eq(init.dat = matrix(runif(10000), nrow = 100))
a <- matrix(dnorm((-49:50)/25), nrow = 100, ncol = 100)
mound <- a+t(a)
smaller.mound <- two.d.heat.eq(n = 3000, init.dat = mound)
plot(smaller.mound)
}
