% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.d.heat.eq.R
\name{one.d.heat.eq}
\alias{one.d.heat.eq}
\title{The Heat Equation Solver
This function takes in all the options possible in order to choose how to solve the heat equation, the boundary conditions, and what language to use.}
\usage{
one.d.heat.eq(init.dat, n = 50000, boundary = c("Neumann",
  "Dirichlet"), alpha = 1, dt = 0.1, eps = 1e-05, cpp = T,
  save.intermediates = F, dx = NULL)
}
\description{
The Heat Equation Solver
This function takes in all the options possible in order to choose how to solve the heat equation, the boundary conditions, and what language to use.
}
